Adds streaming example.
This has a lot of nuances and the API went through a few revisions. Fundamentally there is a conceptual difference between the way the browser streams (with knowledge of the parser and continuously appending to the previously parsed parent) and HTML fragments' approach (streaming top-level nodes independently). There are a few use cases to support here:
1.  Streaming the main HTML document and the user triggers navigation await and then back.
2.  The user navigates and streams the new route, then navigates away and back.

These are both tricky because we have to continue streaming content in the background but append it to the cached route, not the currently displayed one. Use case 1. is particularly tricky because the HTML parser is appending the node and is out of my control. I made use case 1. work by actually caching the `<router-outlet />` element entirely, so subsequent content will continue to append to the cached outlet. If we left the outlet alone and only swapped out its content, the HTML parser would continue appending to that outlet, even when the user has navigated to a different route.

I also realized I was misusing `<slot />` elements. They should only be present in shadow DOM, not light DOM. I made up a `<router-outlet />` tag name (not a real custom element) which serves the same purpose. This is actually good given that if it was a real `<slot />` I don't think it could be removed, cached, and replaced to address use case 1. like I mentioned.

Unfortunately lists get complicated again. I have two examples, `/streaming/` and `/streaming-list/`. The former just streams multiple top-level `<div />` tags. This works just fine and `Router` returns an `AsyncGenerator` to `BaseRouter` which automatically appends them to the outlet as they arrive. The latter is the more realistic use case where a route contains a top-level element (say, a `<ul />` tag) which contains a list of streamable element (`<li />` tags). This doesn't work well with HTML fragments because we only stream top-level elements, so even if some `<li />` tags are delivered early, the user can't see anything until the `</ul>` tag is parsed. I was able to address use case two by client-side rendering the `<ul />` and wrapper `<li />` tags, only streaming the list item contents. This is annoying because it requires forking the server logic for fragments. It renders the `<ul />` tag for real page navigations (because that's how browser streaming actually works), but skips that tag for fragment requests (because that's how HTML fragments requires it). This also means that from `BaseRouter's` perspective there isn't any streaming going on at all, it just receives a single node which magically gets updated over time based on lazy operations from `Router`.