Adds `AsyncGenerator` to the return type of the `route()` method.
This allows streaming-only use cases to be more easily supported. Also regular non-streaming use cases can be implemented as a streaming router which just yields once, so this may be more ergonomic in many instances.

While this could apply to the example router, I'm going to leave it as is to be more clear that it can be simpler in non-streaming cases, rather than making it look like a non-streaming use case still needs an `AsyncGenerator`.